% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_hype}
\alias{bootstrap_hype}
\title{Bootstrap A Hypergraph}
\usage{
bootstrap_hype(hype, n = 1, method = "both")
}
\arguments{
\item{hype}{A hypergraph object.}

\item{n}{The number of bootstrapped hypergraphs required.}

\item{method}{What method to use to calculate the bootstrapped
hypergraphs}
}
\value{
A list of bootstrapped hypergraphs.
}
\description{
Bootstrapping is a common statistical technique used to quantify
uncertainty of calculations. This is an approximation of the
bootstrap algorithm for hypergraphs. Bootstrapping is achieved
by creating a new hypergraph where the vertices, hyperedges or
both have themselves been bootstrapped, achieved using the
\code{"vertex"}, \code{"hyperedge"} or \code{"both"} methods.
}
\examples{
hype <- example_hype()
resamples <- bootstrap_hype(hype, n = 5)
lapply(resamples, incidence_matrix)
}
