% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{cardinality}
\alias{cardinality}
\title{Find The Cardinality Of Hyperedges In A Hypergraph}
\usage{
cardinality(hype)
}
\arguments{
\item{hype}{A hypergraph object.}
}
\value{
A vector of the cardinality of the hyperedges.
}
\description{
The cardinality of a hyperedge is the number of vertices that it contains.
}
\examples{
hype <- example_hype()
cardinality(hype)
}
