% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_graphs.R
\name{connectivity_graph}
\alias{connectivity_graph}
\title{Find The Connectivity Graph Of A Hypergraph}
\usage{
connectivity_graph(hype)
}
\arguments{
\item{hype}{A hypergraph object.}
}
\value{
A graph object representing the hyperedge connectivity graph.
}
\description{
The connectivity graph is a graphical representation of a hypergraph, it has a vertex
for each vertex and hyperedge in the hypergraph. Two hyperedges are connected in the
connectivity graph if they both have a vertex in common, a vertex is connected to  a
hyperedge if the vertex is contained in the hyperedge and no vertices are connected by
edges.
}
\examples{
hype <- example_hype()
g <- connectivity_graph(hype)
print(g)
plot(g)
}
