% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{eigenvector_centrality_factor}
\alias{eigenvector_centrality_factor}
\title{Calculate The Eigenvector Centrality Scaling Factor Of A Hypergraph}
\usage{
eigenvector_centrality_factor(hype)
}
\arguments{
\item{hype}{A hypergraph object}
}
\value{
A number representing the scaling factor relating the value of each node
to the sum of the value of its neighbours
}
\description{
To calculate the eigenvector centrality of a hypergraph, each vertex is assigned
a value that is proportional to the sum of the value of its neighbours. This
function gives the scaling factor relating the value of each node to the sum of
the value of its neighbours.
}
\examples{
h1 <- example_hype()
eigenvector_centrality_factor(h1)
}
