% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{has_real_coef}
\alias{has_real_coef}
\title{Does a Hypergraph Have Real Coefficients}
\usage{
has_real_coef(hype)
}
\arguments{
\item{hype}{A hypergraph object.}
}
\value{
A logical value indicating whether there are real cofficients
associating vertices to hyperedges.
}
\description{
Takes a hypergraph object and returns whether there are real coefficients
associating vertices to hyperedges.
}
\examples{
h <- example_hype()
has_real_coef(h)
}
