% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{hype_dual}
\alias{hype_dual}
\title{Get The Dual Of A Hypergraph}
\usage{
hype_dual(hype)
}
\arguments{
\item{hype}{A hypergraph object.}
}
\value{
A hypergraph object representing the dual of the hypergraph.
}
\description{
The dual of a hypergraph has a vertex for each original hyperedge and a hyperedge
for each original vertex. A vertex is a member of a hyperedge if the original
hyperedge has the original vertex as a member.
}
\examples{
hype <- example_hype()
hype_dual(hype)
}
