% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrices.R
\name{hype_norm_lap_mat}
\alias{hype_norm_lap_mat}
\title{Find the Hyperedge Normalised Laplacian Matrix of a Hypergraph}
\usage{
hype_norm_lap_mat(hype, as_matrix = TRUE)
}
\arguments{
\item{hype}{A hypergraph object}

\item{as_matrix}{Whether to coerce the result to a simple matrix}
}
\value{
The hyperedge normalised laplacian matrix of the hypergraph
}
\description{
As defined by Jurgen Jost and Raffaella Mulas
\doi{10.1016/j.aim.2019.05.025}
}
\examples{
h1 <- example_hype()
hype_norm_lap_mat(h1)
}
