% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{hype_size}
\alias{hype_size}
\title{Get The Size Of A Hypergraph}
\usage{
hype_size(hype)
}
\arguments{
\item{hype}{A hypergraph object}
}
\value{
A number representing the number of hyperedges in a hypergraph
}
\description{
The size of a hypergraph is the number of hyperedges it contains
}
