% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{hyperedge_list}
\alias{hyperedge_list}
\title{Get Hyperedge List}
\usage{
hyperedge_list(hype)
}
\arguments{
\item{hype}{A hypergraph object}
}
\value{
A hyperedge list. See main documentation for more details on its
structure
}
\description{
Take a hypergraph object and return its hyperedge list.
}
\examples{
h <- example_hype()
hyperedge_list(h)
}
