% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{hyperedge_names}
\alias{hyperedge_names}
\title{Get Hyperedge Names}
\usage{
hyperedge_names(hype)
}
\arguments{
\item{hype}{A hypergraph object.}
}
\value{
A vector of strings representing the names of the the hyperedges. If
the hyperedges have no names assocaited with them it will return \code{NULL}
instead.
}
\description{
Takes a hypergraph object and returns the names of the hyperedges.
}
\examples{
h <- example_hype()
hyperedge_names(h)
}
