% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{hyperedge_weights}
\alias{hyperedge_weights}
\title{Get Hyperedge Weights}
\usage{
hyperedge_weights(hype)
}
\arguments{
\item{hype}{A hypergraph object.}
}
\value{
A vector of weights asssociated with the hyperedges. If the are no
weights assicated with the hyperedges then \code{NULL} is returned instead.
}
\description{
Takes a hypergraph object and returns the weights associated with each
hyperedge
}
\examples{
h <- example_hype()
hyperedge_weights(h)
}
