% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{is_directed}
\alias{is_directed}
\title{Is a Hypergraph Directed}
\usage{
is_directed(hype)
}
\arguments{
\item{hype}{A hyeprgraph object.}
}
\value{
A logical value indicating whether the hyperedges are directed.
}
\description{
Takes a hypergraph object and returns whether the hyperedges are directed.
}
\examples{
h <- example_hype()
is_directed(h)
}
