% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{is_oriented}
\alias{is_oriented}
\title{Is a Hypergraph Oriented}
\usage{
is_oriented(hype)
}
\arguments{
\item{hype}{A hypergraph object.}
}
\value{
A logical value indicating whether the hyperedges are oriented.
}
\description{
Takes a hypergraph object and returns whether the hyperedges are oriented.
}
\examples{
h <- example_hype()
is_oriented(h)
}
