% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{is_weighted}
\alias{is_weighted}
\title{Is a Hypergraph Weighted}
\usage{
is_weighted(hype)
}
\arguments{
\item{hype}{A hypergraph object.}
}
\value{
A logical value indicating whether the hypergraph has weights
associated with its vertices or hyperedges.
}
\description{
Takes a hypergraph object and returns whether a hypergraph has weights
associated with its vertices or hyperedges.
}
\examples{
h <- example_hype()
is_weighted(h)
}
