% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrices.R
\name{laplacian_matrix}
\alias{laplacian_matrix}
\title{Find the Laplacian Matrix of a Hypergraph}
\usage{
laplacian_matrix(hype)
}
\arguments{
\item{hype}{A hypergraph object}
}
\value{
The laplacian matrix of the hypergraph
}
\description{
Find the Laplacian Matrix of a Hypergraph
}
\examples{
h1 <- example_hype()
laplacian_matrix(h1)
}
