% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{max_cardinality}
\alias{max_cardinality}
\title{Find The Maximum Cardinality Of A Hyperedge In A Hypergraph}
\usage{
max_cardinality(hype)
}
\arguments{
\item{hype}{A hypergraph object.}
}
\value{
The value of the maximum cardinality of a hyperedge in the hypergraph.
}
\description{
The cardinality of a hyperedge is the number of vertices that it contains.
}
\examples{
hype <- example_hype()
max_cardinality(hype)
}
