% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{min_cardinality}
\alias{min_cardinality}
\title{Find The Minimum Cardinality Of A Hyperedge In A Hypergraph}
\usage{
min_cardinality(hype)
}
\arguments{
\item{hype}{A hypergraph object.}
}
\value{
The value of the minimum cardinality of a hyperedge in the hypergraph.
}
\description{
The cardinality of a hyperedge is the number of vertices that it contains.
}
\examples{
hype <- example_hype()
min_cardinality(hype)
}
