% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{min_degree}
\alias{min_degree}
\title{Find The Minimum Degree Of A Hypergraph}
\usage{
min_degree(hype, method = NA)
}
\arguments{
\item{hype}{A hypergraph object.}

\item{method}{The method by which to calculate the degree, see \code{degree} help
documentation for more information.}
}
\value{
The maximum value for the chosen degree among vertices of the hypergraph.
}
\description{
The degree of a vertex is a way of expressing how many connections there
are from a vertex to the rest of the hypergraph. See the \code{degree} help
documentation for more details.
}
\examples{
h <- example_hype()
min_degree(h)
}
