% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{partial_hype}
\alias{partial_hype}
\title{Generate A Partial Hypergraph}
\usage{
partial_hype(hype, hyperedges)
}
\arguments{
\item{hype}{A hypergraph object.}

\item{hyperedges}{A vector of the names of the hyperedges to be used to
induce the partial hypergraph.}
}
\value{
A hypergraph object of the partial hypergraph.
}
\description{
A partial hypergraph can be induced from a set of hyperedges. The partial
hypergraph has all of the original vertices, but only the hyperedges used
to induce it.
}
\examples{
hype <- example_hype()
partial_hype(hype, c("h1", "h2"))
}
