% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrices.R
\name{pseudo_invert}
\alias{pseudo_invert}
\title{Pseudo-Invert a Vector}
\usage{
pseudo_invert(vec)
}
\arguments{
\item{vec}{A vector of numbers}
}
\value{
A vector of pseudo-inverted numbers
}
\description{
Pseudoinversion is where a vector has each non-zero element inverted and each zero
element remains untouched. This is useful for pseudoinverting matrices that only have
non-zero entries on the leading diagonal.
}
