% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperpaths.R
\name{shortest_hyperpaths}
\alias{shortest_hyperpaths}
\title{Find The Shortest Hyperpaths Between Two Vertices}
\usage{
shortest_hyperpaths(hype, from, to)
}
\arguments{
\item{hype}{A hypergraph object.}

\item{from}{The vertex that is the start of the hyperpath.}

\item{to}{The vertex that is the end of the hyperpath,}
}
\value{
A list of shortest hyperpaths between the given vertices.
}
\description{
A hyperpath is a set of hyperedges such that each consecutive pair of hyperedges
contain at least vertex in common. A shortest hyperpath between two vertices is
the smallest set of hyperedges that form a hyperpath such that one vertex is in
the first hyperpath and the other vertex is in the last hyperpath.
}
\examples{
hype <- example_hype()
shortest_hyperpaths(hype, "v1", "v4")
}
