% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamples.R
\name{shuffle_hype}
\alias{shuffle_hype}
\title{Shuffle A Hypergraph}
\usage{
shuffle_hype(hype, n = 1, method = "hyperedge")
}
\arguments{
\item{hype}{A hypergraph object.}

\item{n}{The number of shuffled hypergraphs to calculate.}

\item{method}{The method to use to shuffle the hypergraph.}
}
\value{
A list of shuffled hypergraphs.
}
\description{
A hypergraph can be shuffled to slightly perturb its structure. These shuffled
hypergraphs can then be used to estimate the uncertainty of calculations on
the original hypergraph.
}
\details{
Two methods are used to shuffle a hypergraph, the \code{"vertex"} method keeps the
degree of each vertex the same, randomly reassigning the hyperedges they are
members of. The \code{"hyperedge"} method keeps the cardinality of each hyperedge
the same, randomly reassigning the vertices that are members.
}
\examples{
hype <- example_hype()
resamples <- shuffle_hype(hype, n = 5)
lapply(resamples, incidence_matrix)
}
