% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_calcs.R
\name{spectra}
\alias{spectra}
\title{Find the Spectra of a Hypergraph}
\usage{
spectra(hype, matrix = "laplacian")
}
\arguments{
\item{hype}{A hypergraph object}

\item{matrix}{The matrix to calculate the spectra with respect to. Out of
\code{"laplacian"}, \code{"adjacency"}, \code{"vert_norm_lap_mat"} and \code{"hype_norm_lap_mat"}}
}
\value{
The eigen decomposition of the given matrix of the given hypergraph
}
\description{
Find the Spectra of a Hypergraph
}
\examples{
h <- example_hype()
spectra(h)
}
