% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_calcs.R
\name{spectral_distance_disc}
\alias{spectral_distance_disc}
\title{Find the Spectral Distance From the Fully Disconnected Hypergraph}
\usage{
spectral_distance_disc(hype, matrix = "vert_norm_lap_mat")
}
\arguments{
\item{hype}{A hypergraph object}

\item{matrix}{The matrix to calculate the spectra with respect to. Out of
\code{"vert_norm_lap_mat"} and \code{"hype_norm_lap_mat"}}
}
\value{
The spectral distance from the disconnected hypergraph
}
\description{
Find the Spectral Distance From the Fully Disconnected Hypergraph
}
\examples{
h <- example_hype()
spectral_distance_disc(h)
}
