% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_graphs.R
\name{support_graph}
\alias{support_graph}
\title{Find The Support Graph Of A Hypergraph}
\usage{
support_graph(hype, simple = TRUE)
}
\arguments{
\item{hype}{A hypergraph object.}

\item{simple}{Whether a simplified support graph should be created.}
}
\value{
The support graph of the hypergraph
}
\description{
The support graph of a hypergraph is a graph that has a vertex for every vertex
in the hypergraph. Two vertices are connected in the support graph if there is a
hyperedge that connects them.
}
\details{
If \code{simple} is set to \code{FALSE} then for unweighted hypergraphs without real coefficients
the support graph has an edge connected vertices for each hyperedge connecting them
in the hypergraph.
}
\examples{
hype <- example_hype()
g <- support_graph(hype)
print(g)
plot(g)
}
