% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrices.R
\name{vert_norm_lap_mat}
\alias{vert_norm_lap_mat}
\title{Find the Vertex Normalised Laplacian Matrix of a Hypergraph}
\usage{
vert_norm_lap_mat(hype)
}
\arguments{
\item{hype}{A hypergraph object}
}
\value{
The vertex normalised laplacian matrix of the hypergraph
}
\description{
Find the Vertex Normalised Laplacian Matrix of a Hypergraph
}
