% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{vertex_names}
\alias{vertex_names}
\title{Get Vertex Names}
\usage{
vertex_names(hype)
}
\arguments{
\item{hype}{A hypergraph object.}
}
\value{
A vector of strings of vertex names
}
\description{
Takes a hypergraph object and returns the names of its vertices.
}
\examples{
h <- example_hype()
vertex_names(h)
}
