% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_setters.R
\name{vertex_weights}
\alias{vertex_weights}
\title{Get Vertex Weights}
\usage{
vertex_weights(hype)
}
\arguments{
\item{hype}{A hypergraph object.}
}
\value{
A vector of weights associated with each vertex. If the hypergraph
has no weights associated with its vertices it will return \code{NULL} instead.
}
\description{
Takes a hypergraph object and returns the weights associated with its
vertices.
}
\examples{
h <- example_hype()
vertex_weights(h)
}
