% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrices.R
\name{pseudo_invert}
\alias{pseudo_invert}
\title{Pseudo-Invert a Vector}
\usage{
pseudo_invert(vec)
}
\arguments{
\item{vec}{A vector of numbers}
}
\value{
A vector of pseudo-inverted numbers
}
\description{
Pseudo-Invert a Vector
}
