% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrices.R
\name{vert_norm_lap_mat}
\alias{vert_norm_lap_mat}
\title{Find the Vertex Normalised Laplacian Matrix of a Hypergraph}
\usage{
vert_norm_lap_mat(hype)
}
\arguments{
\item{hype}{A hypergraph object}
}
\value{
The vertex normalised laplacian matrix of the hypergraph
}
\description{
As defined by Jurgen Jost and Raffaella Mulas
\doi{10.1016/j.aim.2019.05.025}
}
\examples{
h1 <- example_hype()
vert_norm_lap_mat(h1)
}
